Make filters reusable
The new homepage will duplicate the filters from the facilities search
tab. Splits the filters into separate components as necessary, and adds
a new StyledSelect component, to keep the code DRY.
